Make Docker image actually publishable & usable as-is for reporting
The tool is now built inside the container.
And if we publish the Docker image in the future, one could run it
right away using the "embedded"/built fatjar.

Side note: for now, I left the base image as `maven:xyz` on purpose without
using multi-stage builds, so that possibly users can use this too to
rebuild the tool locally (by bind mounting the sources).
I might change this later, but given this is mostly a dev tool, there's
no clear incentive here IMO.